Move tests to a manual tests file Add sustained load test
