+++ tests/heavy_tests.rs
@@ -1,25 +1,17 @@
-use futures_util::{
-    future::{join_all, AbortHandle, Abortable},
-    sink, stream,
-    stream::FuturesUnordered,
-    StreamExt,
-};
+use futures_util::{future::join_all, sink, stream::FuturesUnordered, StreamExt};
 use std::{
-    mem,
-    net::SocketAddr,
     sync::Arc,
     time::{Duration, Instant},
 };
-use temporal_client::{GetWorkflowResultOpts, WfClientExt, WorkflowClientTrait, WorkflowOptions};
+use temporal_client::{WfClientExt, WorkflowClientTrait, WorkflowOptions};
 use temporal_sdk::{ActContext, ActivityOptions, WfContext, WorkflowResult};
-use temporal_sdk_core::{CoreRuntime, ResourceBasedTuner, ResourceSlotOptions};
-use temporal_sdk_core_api::{telemetry::PrometheusExporterOptionsBuilder, worker::PollerBehavior};
+use temporal_sdk_core::{ResourceBasedTuner, ResourceSlotOptions};
+use temporal_sdk_core_api::worker::PollerBehavior;
 use temporal_sdk_core_protos::{
     coresdk::{AsJsonPayloadExt, workflow_commands::ActivityCancellationType},
     temporal::api::enums::v1::WorkflowIdReusePolicy,
 };
-use temporal_sdk_core_test_utils::{prom_metrics, workflows::la_problem_workflow, CoreWfStarter};
-use tracing::info;
+use temporal_sdk_core_test_utils::{workflows::la_problem_workflow, CoreWfStarter};
 
 mod fuzzy_workflow;
 
@@ -368,144 +360,4 @@ async fn can_paginate_long_history() {
         }
     });
     worker.run_until_done().await.unwrap();
-}
-
-// TODO: Multithread via macro doesn't set trace subscriber properly
-// #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-#[tokio::test]
-async fn poller_load() {
-    const SIGNAME: &str = "signame";
-    let num_workflows = 1_000;
-    let wf_name = "poller_load";
-    let (telemopts, addr, _aborter) = prom_metrics(Some(
-        PrometheusExporterOptionsBuilder::default()
-            .socket_addr(SocketAddr::V4("0.0.0.0:9999".parse().unwrap()))
-            .build()
-            .unwrap(),
-    ));
-    let rt = CoreRuntime::new_assume_tokio(telemopts).unwrap();
-    let mut starter = CoreWfStarter::new_with_runtime("poller_load", rt);
-    starter
-        .worker_config
-        .max_cached_workflows(5000_usize)
-        .max_outstanding_workflow_tasks(1000_usize)
-        .workflow_task_poller_behavior(PollerBehavior::Autoscaling {
-            minimum: 1,
-            maximum: 200,
-            initial: 5,
-        })
-        .no_remote_activities(true);
-    let mut worker = starter.worker().await;
-    let submitter = worker.get_submitter_handle();
-    worker.register_wf(wf_name.to_owned(), |ctx: WfContext| async move {
-        let sigchan = ctx.make_signal_channel(SIGNAME).map(Ok);
-        let drained_fut = sigchan.forward(sink::drain());
-
-        let real_stuff = async move {
-            for _ in 0..10 {
-                ctx.timer(Duration::from_millis(1)).await;
-            }
-        };
-        tokio::select! {
-            _ = drained_fut => {}
-            _ = real_stuff => {}
-        }
-
-        Ok(().into())
-    });
-    let client = starter.get_client().await;
-
-    info!("Prom bound to {:?}", addr);
-
-    let mut workflow_handles = vec![];
-    info!("Starting workflows...");
-    for i in 0..num_workflows {
-        let wfid = format!("{wf_name}_{i}");
-        let rid = worker
-            .submit_wf(
-                wfid.clone(),
-                wf_name.to_owned(),
-                vec![],
-                WorkflowOptions {
-                    execution_timeout: Some(Duration::from_secs(100)),
-                    ..Default::default()
-                },
-            )
-            .await
-            .unwrap();
-        workflow_handles.push(client.get_untyped_workflow_handle(wfid, rid));
-    }
-    info!("Done starting workflows");
-    let start_processing = Instant::now();
-
-    let (ah, abort_reg) = AbortHandle::new_pair();
-    let all_workflows_are_done = async {
-        stream::iter(mem::take(&mut workflow_handles))
-            .for_each_concurrent(25, |handle| async move {
-                let _ = handle
-                    .get_workflow_result(GetWorkflowResultOpts::default())
-                    .await;
-            })
-            .await;
-        info!("Initial load ran for {:?}", start_processing.elapsed());
-        // TODO: Maybe send signals for round two
-        ah.abort();
-        // Wait a minute for poller count to drop
-        tokio::time::sleep(Duration::from_secs(60)).await;
-        // round two
-        let start_processing = Instant::now();
-        for i in 0..500 {
-            let wfid = format!("{wf_name}_2_{i}");
-            let rid = submitter
-                .submit_wf(
-                    wfid.clone(),
-                    wf_name.to_owned(),
-                    vec![],
-                    WorkflowOptions {
-                        execution_timeout: Some(Duration::from_secs(100)),
-                        ..Default::default()
-                    },
-                )
-                .await
-                .unwrap();
-            workflow_handles.push(client.get_untyped_workflow_handle(wfid, rid));
-        }
-        stream::iter(workflow_handles)
-            .for_each_concurrent(25, |handle| async move {
-                let _ = handle
-                    .get_workflow_result(GetWorkflowResultOpts::default())
-                    .await;
-            })
-            .await;
-        info!("Second load ran for {:?}", start_processing.elapsed());
-
-        info!("Done. Go look at yer metrics, must interrupt to end.");
-    };
-
-    let sig_sender = Abortable::new(
-        async {
-            loop {
-                let sends: FuturesUnordered<_> = (0..num_workflows)
-                    .map(|i| {
-                        client.signal_workflow_execution(
-                            format!("{wf_name}_{i}"),
-                            "".to_string(),
-                            SIGNAME.to_string(),
-                            None,
-                            None,
-                        )
-                    })
-                    .collect();
-                sends
-                    .map(|_| Ok(()))
-                    .forward(sink::drain())
-                    .await
-                    .expect("Sending signals works");
-                tokio::time::sleep(Duration::from_secs(2)).await;
-            }
-        },
-        abort_reg,
-    );
-    let (runres, _, _) = tokio::join!(worker.inner_mut().run(), sig_sender, all_workflows_are_done);
-    runres.unwrap();
 }
